requirements.txt
pandas
openpyxl
mysql-connector-python
ttkbootstrap
